requirements.txt
google-generativeai
feedparser
requests
python-dotenv
tweepy
gspread
google-api-python-client
